Update Emily_Zhu_Project_1.pde
I tried to make mine less literal by adding other stuff and modifying it so that it wasn't just a tree growing. I also switched the adjective which made it a little easier to make it less literal because natural was a bit difficult in finding shapes and movement that corresponded without making it completely literal.